move adapter classes to its own files
