+++ lib/design_patterns/adapter/contracts/service_one.dart
@@ -1,0 +1,5 @@
+import '../customer.dart';
+
+abstract class ServiceOne {
+  Customer makeCustomer();
+}
