+++ lib/design_patterns/adapter/contracts/service_two.dart
@@ -1,0 +1,5 @@
+import '../customer.dart';
+
+abstract class ServiceTwo {
+  Customer findCustomer();
+}
