+++ lib/design_patterns/adapter/customer.dart
@@ -1,0 +1,6 @@
+class Customer {
+  final String fullName;
+  final int age;
+
+  Customer(this.fullName, this.age);
+}
